requirements.txt
kubernetes
kafka-python
requests
textual
paramiko
Flask==2.3.2
requests==2.30.0
